pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "streamlit>=1.43.1",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "altair>=5.0.0",
]
